- fix isAuthenticated flag - renamed EmailVerification to RegisterEmailForm - cannot run lint
